requirements.txt
wsgiref==0.1.2
Django==1.8.2
Jinja2==2.7.3
Markdown==2.6.2
MarkupSafe==0.23
Pillow==2.8.1
PyJWT==0.4.1
PyYAML==3.11
Pygments==2.0.2
South==1.0.2
Wand==0.3.7
django-angular==0.7.14
django-appconf==1.0.1
django-bower==5.0.4
django-bower-app==0.5
django-compressor==1.5
django-filter==0.9.2
djangorestframework==3.1.1
httpie==0.9.2
inflection==0.3.0
npm==0.1.1
oauthlib==0.7.2
optional-django==0.1.0
pep8==1.5.7
psycopg2==2.5.2
python-openid==2.2.5
python-social-auth==0.2.2
requests==2.6.0
requests-oauthlib==0.4.2
rest-framework-ember==1.3.1
six==1.9.0
virtualenv==12.0.7
django-extensions==1.5.5
Werkzeug==0.10.4
django-cross-origin==0.9.0
django-cities-light==3.1.2
dj_static

